Only add 'drafts' tag for draft posts if it isn't already a tag
In looking to find a fix for #61 I discovered this little check that should be
done. However, the original bug defined in #61 was not able to be reproduced. So
not sure what was up there. This is better but does not actually solve the
problem if the problem still even exists.