Made a general "default data" map; gave Posts default layout
The `DefaultPermalinksMap` was a gateway to realizing bundles might want to have
a lot of default things. Rather than creating, say, a `DefaultLayoutMap` as
well, anything that looks like this can be handled with a simple collection of
key => value pairs.

This will be a BC break for anyone who has ever written a post w/o a layout
intentionally.

Closes #50